requirements.txt
Django==2.2.5
gunicorn==19.9.0
django-heroku==0.3.1
django-debug-toolbar==2.0
docutils==0.15.2
